Task 1.5.2 Moved Node class to IntLinkedList.java
