+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -2,6 +2,8 @@ package edu.kis.vh.nursery.list;
 
 public class IntLinkedList {
 
+    private static final int ERROR = -1;
+
     public Node getLast() {
         return last;
     }
@@ -30,13 +32,13 @@ public class IntLinkedList {
 
     public int top() {
         if (isEmpty())
-            return -1;
+            return ERROR;
         return last.getValue();
     }
 
     public int pop() {
         if (isEmpty())
-            return -1;
+            return ERROR;
         int ret = last.getValue();
         last = last.getPrev();
         return ret;
@@ -46,4 +48,34 @@ public class IntLinkedList {
         return i;
     }
 
+}
+class Node {
+
+    private int value;
+    private Node prev;
+    private Node next;
+
+    public Node(int i) {
+        value = i;
+    }
+
+    public int getValue() {
+        return value;
+    }
+
+    public Node getPrev() {
+        return prev;
+    }
+
+    public void setPrev(Node prev) {
+        this.prev = prev;
+    }
+
+    public Node getNext() {
+        return next;
+    }
+
+    public void setNext(Node next) {
+        this.next = next;
+    }
 }
